requirements.txt
pandas==1.3.5
python-dotenv==0.19.2
SQLAlchemy==1.3.24
psycopg2-binary==2.8.6